+++ CloudSeed/AudioLib/ValueTables.h
@@ -11,28 +11,28 @@ namespace AudioLib
 	public:
 		static const int TableSize = 40001;
 
-		static DSY_SDRAM_BSS double Sqrt[TableSize];
-		static DSY_SDRAM_BSS double Sqrt3[TableSize];
-		static DSY_SDRAM_BSS double Pow1_5[TableSize];
-		static DSY_SDRAM_BSS double Pow2[TableSize];
-		static DSY_SDRAM_BSS double Pow3[TableSize];
-		static DSY_SDRAM_BSS double Pow4[TableSize];
-		static DSY_SDRAM_BSS double x2Pow3[TableSize];
+		static DSY_SDRAM_BSS float Sqrt[TableSize];
+		static DSY_SDRAM_BSS float Sqrt3[TableSize];
+		static DSY_SDRAM_BSS float Pow1_5[TableSize];
+		static DSY_SDRAM_BSS float Pow2[TableSize];
+		static DSY_SDRAM_BSS float Pow3[TableSize];
+		static DSY_SDRAM_BSS float Pow4[TableSize];
+		static DSY_SDRAM_BSS float x2Pow3[TableSize];
 
-		// octave response. value double every step (2,3,4,5 or 6 steps)
-		static DSY_SDRAM_BSS double Response2Oct[TableSize];
-		static DSY_SDRAM_BSS double Response3Oct[TableSize];
-		static DSY_SDRAM_BSS double Response4Oct[TableSize];
-		static DSY_SDRAM_BSS double Response5Oct[TableSize];
-		static DSY_SDRAM_BSS double Response6Oct[TableSize];
+		// octave response. value float every step (2,3,4,5 or 6 steps)
+		static DSY_SDRAM_BSS float Response2Oct[TableSize];
+		static DSY_SDRAM_BSS float Response3Oct[TableSize];
+		static DSY_SDRAM_BSS float Response4Oct[TableSize];
+		static DSY_SDRAM_BSS float Response5Oct[TableSize];
+		static DSY_SDRAM_BSS float Response6Oct[TableSize];
 
 		// decade response, value multiplies by 10 every step
-		static DSY_SDRAM_BSS double Response2Dec[TableSize];
-		static DSY_SDRAM_BSS double Response3Dec[TableSize];
-		static DSY_SDRAM_BSS double Response4Dec[TableSize];
+		static DSY_SDRAM_BSS float Response2Dec[TableSize];
+		static DSY_SDRAM_BSS float Response3Dec[TableSize];
+		static DSY_SDRAM_BSS float Response4Dec[TableSize];
 
 		static void Init();
-		static double Get(double index, double* table);
+		static float Get(float index, float* table);
 	};
 }
