+++ CloudSeed/Utils.h
@@ -11,18 +11,18 @@ namespace CloudSeed
 	{
 	public:
 
-		static inline void ZeroBuffer(double* buffer, int len)
+		static inline void ZeroBuffer(float* buffer, int len)
 		{
 			for (int i = 0; i < len; i++)
 				buffer[i] = 0.0;
 		}
 
-		static inline void Copy(double* source, double* dest, int len)
+		static inline void Copy(float* source, float* dest, int len)
 		{
-			std::memcpy(dest, source, len * sizeof(double));
+			std::memcpy(dest, source, len * sizeof(float));
 		}
 
-		static inline void Gain(double* buffer, double gain, int len)
+		static inline void Gain(float* buffer, float gain, int len)
 		{
 			for (int i = 0; i < len; i++)
 			{
@@ -34,12 +34,12 @@ namespace CloudSeed
 		// undersampling: if set to 1, perfroms no effect, if set to 2, will undersample to 1/2 samplerate, etc...
 		// sampleResolution: if set to 32, will use 2^32 steps, if set to 8, will resude to 2^8 = 256 steps
 		// Currently Unused
-		static inline void BitcrushAndReduce(double* bufferIn, double* bufferOut, int len, int undersampling, int sampleResolution)
+		static inline void BitcrushAndReduce(float* bufferIn, float* bufferOut, int len, int undersampling, int sampleResolution)
 		{
-			double sampleSteps = std::pow(2, sampleResolution);
-			double inverseSteps = 1.0 / sampleSteps;
+			float sampleSteps = std::pow(2, sampleResolution);
+			float inverseSteps = 1.0 / sampleSteps;
 
-			double sample = 0.0;
+			float sample = 0.0;
 
 			for (int i = 0; i < len; i++)
 			{
@@ -51,13 +51,13 @@ namespace CloudSeed
 		}
 
 		template<typename T>
-		static double DB2gain(T input)
+		static float DB2gain(T input)
 		{
 			return std::pow(10, input / 20.0);
 		}
 
 		template<typename T>
-		static double Gain2DB(T input)
+		static float Gain2DB(T input)
 		{
 			if (input < 0.0000001)
 				return -100000;
